requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
numpy==1.26.0
opencv-python==4.8.0.76
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
